move relay related code
